Fix rendering of indented lines in text/plain parts
Mail has a rendering bug which drops one leading space from indented
lines in text/plain messages. As well as interfering with column
alignment, this also corrupts inline patches which could otherwise be
copied and pasted directly into patch or git apply.

When converting the text/plain part to HTML for display in a WebView,
all but the first leading space is rendered as '&nbsp;' so, for example,
a line '  foo' becomes '<BR> &nbsp;foo'. However, WebKit will then
display this as ' foo' not '  foo'.

To work around this, swizzle the decodeTextPlainWithContext: selector of
MCMimePart to translate a space following a '<BR>' (or at the start of
the rendered fragment) into '&nbsp;' just before returning the HTML for
rendering. Mail always translates line-breaks to '<BR>', not '<br>' or
'<BR />'. A case-sensitive literal match is therefore used for speed,
minimising the impact on UI latency.