Add "FixAttribution" user preference
The arachsys/mailwrap Apple Mail plugin on GitHub includes a user
preference to toggle on or off whether the plugin will fix the
attribution on replies. Mailflow lacks this user preference. This change
adds the same support for the "FixAttribution" user preference in
MailWrap to MailFlow.

The same sematics in MailWrap applies here. This information from the
MailWrap README also applies to this feature added to MailFlow:

  defaults write com.apple.mail MailWrap -dict-add FixAttribution -bool false
  defaults write com.apple.mail MailWrap -dict-add FixAttribution -bool true
    - configure MailWrap to strip the verbose date and time information
      from the attribution line when composing a reply. The default is on.

Signed-off-by: Paul Mather <paul.mather@vt.edu>