Add FlowWidth setting to control the width of format=flowed text
Allow the default width of format=flowed text to be reconfigured from
the default 76 characters. Setting the width to zero will globally
disable format=flowed but leave the other features of MailFlow intact,
including the restriction on unnecessary quoted-printable.

Log successful startup to the system log to match MailWrap, for easier
debugging.

Document the FixAttribution and FlowWidth configuration options.